Serialize numbers in json value to int64

+++ src/Serializer.fs
@@ -73,7 +73,7 @@ module Serialize =
 
         let rec toSerializableJson: JsonValue -> obj = function
             | JsonValue.String string -> string :> obj
-            | JsonValue.Number number -> number |> int :> obj
+            | JsonValue.Number number -> number |> int64 :> obj
             | JsonValue.Float float -> float :> obj
             | JsonValue.Boolean bool -> bool :> obj
             | JsonValue.Record properties ->
@@ -89,7 +89,7 @@ module Serialize =
 
         let rec toSerializableJsonIgnoringNullsInRecord: JsonValue -> obj = function
             | JsonValue.String string -> string :> obj
-            | JsonValue.Number number -> number |> int :> obj
+            | JsonValue.Number number -> number |> int64 :> obj
             | JsonValue.Float float -> float :> obj
             | JsonValue.Boolean bool -> bool :> obj
             | JsonValue.Record properties ->
